fix: Resolve all remaining build errors from merged PRs
Completed Fixes:
1. CaseIterable conformance - Removed redundant extensions
2. ShareSheet redeclaration - Removed duplicate, kept cross-platform version
3. LocalizationKey.customKey - Added case and handling
4. ModularCanvas compatibility - Added queryDocuments overload
5. Main Actor issues - Made stopSession() nonisolated
6. iOS HSplitView compatibility - Replaced with HStack in all views
   - CollaborativeReviewView
   - SlideEditorView
   - ScriptEditorView
   - ModularCanvasView
7. Validation fixes - Changed let to var, fixed return types
8. Type fixes - Date formatting, Double to Int conversions
9. IntelligentCapture fixes:
   - Added await to queueForOfflineProcessing calls
   - Changed allActions from var to let
   - Fixed optional binding for non-optional imageData
   - Fixed Citation struct usage to match actual definition

Build Status: ✅ BUILD SUCCEEDED with zero errors (warnings only)
Xcode: 15.2 | iOS Simulator: iPhone 17 Pro

+++ AFHAM/Features/DocsWorkspace/ModularCanvas/LocalizationView.swift
@@ -371,7 +371,7 @@ struct LocalizationView: View {
 
         // Check if translation is complete
         results.localizationComplete = !layer.translatedContent.isEmpty &&
-                                       layer.translatedContent.count > layer.sourceContent.count * 0.5
+                                       layer.translatedContent.count > Int(Double(layer.sourceContent.count) * 0.5)
 
         // Check tone compliance (simple heuristic)
         results.toneCompliance = true
